Refactored user router
