+++ users.test.js
@@ -1,7 +1,7 @@
-const { registerUser } = require("./routes/users");
+const userRegister = require("./routes/userRegister");
 const bcrypt = require('bcrypt');
-const User = require("./classes/user");
-const Token = require("./classes/token");
+const User = require("./classes/User");
+const Token = require("./classes/Token");
 
 jest.mock("./classes/user");
 jest.mock("./classes/token");
@@ -26,7 +26,7 @@ const mockResponse = {
 
 it("Should return a status code of 400 when properties are missing", async () => {
 
-    await registerUser(incompleteMockRequest, mockResponse);
+    await userRegister(incompleteMockRequest, mockResponse);
 
     expect(mockResponse.status).toHaveBeenCalledWith(400);
     expect(mockResponse.send).toHaveBeenCalledTimes(1);
@@ -39,7 +39,7 @@ it("Should return a status code of 400 when user exists", async () => {
         password: "testingHashed"
     });
 
-    await registerUser(mockRequest, mockResponse);
+    await userRegister(mockRequest, mockResponse);
 
     expect(mockResponse.status).toHaveBeenCalledWith(400);
     expect(mockResponse.send).toHaveBeenCalledTimes(1);
@@ -55,7 +55,7 @@ it("Should return a status code of 200 when new user is created", async () => {
 
     bcrypt.hash.mockResolvedValueOnce("testingHashed");
 
-    await registerUser(mockRequest, mockResponse);
+    await userRegister(mockRequest, mockResponse);
 
     expect(mockResponse.status).toHaveBeenCalledWith(201);
     expect(mockResponse.send).toHaveBeenCalledTimes(1);
